pyproject.toml
[project]
name = "finlab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "edgartools>=5.10.2",
    "fastapi>=0.128.0",
    "fastembed>=0.7.4",
    "groq>=1.0.0",
    "guardrails-ai>=0.8.0",
    "h>=0.99",
    "hdbscan>=0.8.41",
    "instructor>=1.14.5",
    "langfuse>=3.14.1",
    "openai>=2.16.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "sentence-transformers>=5.2.0",
    "trafilatura>=2.0.0",
    "transformers>=4.57.3",
    "uvicorn>=0.40.0",
    "yfinance>=1.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
